Breaking framework helpers to standalone files
