+++ composer.json
@@ -75,6 +75,11 @@
     "autoload": {
         "files": [
             "src/mantle/framework/helpers.php",
+            "src/mantle/application/autoload.php",
+            "src/mantle/assets/autoload.php",
+            "src/mantle/cache/autoload.php",
+            "src/mantle/config/autoload.php",
+            "src/mantle/http/autoload.php",
             "src/mantle/queue/autoload.php",
             "src/mantle/support/autoload.php",
             "src/mantle/testing/autoload.php"
