+++ src/mantle/cache/composer.json
@@ -26,6 +26,11 @@
             "email": "info@alley.co"
         }
     ],
+    "autoload": {
+        "files": [
+            "autoload.php"
+        ]
+    },
     "minimum-stability": "dev",
     "config": {
         "sort-packages": true
