+++ src/mantle/framework/bootstrap/class-load-configuration.php
@@ -8,7 +8,7 @@
 namespace Mantle\Framework\Bootstrap;
 
 use Exception;
-use Mantle\Framework\Application;
+use Mantle\Application\Application;
 use Mantle\Config\Repository;
 use Mantle\Contracts\Config\Repository as Repository_Contract;
 use Mantle\Support\Arr;
