+++ src/mantle/framework/bootstrap/class-load-environment-variables.php
@@ -10,7 +10,7 @@ namespace Mantle\Framework\Bootstrap;
 use Dotenv\Dotenv;
 use Dotenv\Exception\InvalidFileException;
 use Mantle\Framework\Console\Kernel;
-use Mantle\Framework\Application;
+use Mantle\Application\Application;
 use Mantle\Support\Environment;
 
 /**
