+++ src/mantle/framework/bootstrap/class-register-facades.php
@@ -8,7 +8,7 @@
 namespace Mantle\Framework\Bootstrap;
 
 use Mantle\Framework\Alias_Loader;
-use Mantle\Framework\Application;
+use Mantle\Application\Application;
 use Mantle\Facade\Facade;
 use Mantle\Framework\Package_Manifest;
