+++ src/mantle/http/composer.json
@@ -31,6 +31,11 @@
             "email": "info@alley.co"
         }
     ],
+    "autoload": {
+        "files": [
+            "autoload.php"
+        ]
+    },
     "minimum-stability": "dev",
     "config": {
         "sort-packages": true
